Mostly feature complete

+++ src/actions/article.js
@@ -1,3 +1,6 @@
+import { modal } from './modal.js';
+import { displayAlert } from './alerts.js';
+
 export const emptyArticle = {
   title: '',
   description: '',
@@ -47,12 +50,15 @@ export const loadArticle = async ({ state, set }, { slug, loadComments, loadProf
   }
 };
 
-export const postArticle = async ({ data, state, set }) => {
+export const postArticle = async ({ data, state, set, dispatch }) => {
   const { api, article } = state;
   const { slug } = article;
 
   // When a new article is created, the slug is undefined
   const apiFn = slug ? api.Articles.update : api.Articles.create;
+  const msg = slug
+    ? 'Article was updated successfully!'
+    : 'New article was posted sucessfully!';
 
   await set({ busy: true });
 
@@ -62,6 +68,19 @@ export const postArticle = async ({ data, state, set }) => {
     await set({
       article: updated,
       busy: false,
+    });
+
+    await dispatch(displayAlert, {
+      type: "success",
+      timeout: 10000,
+      text: (
+        <span>
+          {msg} Keep editing, or&nbsp;
+          <a className="alert-link" href={`/article/${updated.slug}`}>
+            view it here.
+          </a>
+        </span>
+      )
     });
 
     data.navigate(`/editor/${updated.slug}`);
@@ -74,13 +93,37 @@ export const postArticle = async ({ data, state, set }) => {
   }
 };
 
-export const deleteArticle = async ({ data, state, set }, slug) => {
+export const deleteArticle = async ({ data, state, set, dispatch }, slug) => {
+  const { article: { title } } = state;
+
+  const yesButton = <button className="btn btn-primary">Yes!</button>;
+
+  const result = await dispatch(modal, {
+    title: "Confirm deleting article",
+    content: `Are you sure you want to delete the article titled ${title}?`,
+    buttons: [
+      yesButton,
+      <button className="btn btn-secondary">No, keep it for now</button>
+    ]
+  });
+
+  if (result !== yesButton) {
+    return;
+  }
+
   await set({ busy: true });
 
   try {
     // If no error is thrown, request was successful.
     await state.api.Articles.delete(slug);
 
+    await dispatch(displayAlert, {
+      type: "success",
+      text: `Article titled ${title} was deleted successfully.`
+    });
+
+    // Deletion is allowed only in the full Arcticle view, and
+    // after the article is deleted there is nothing to display.
     data.navigate('/');
   }
   catch (e) {
@@ -88,10 +131,15 @@ export const deleteArticle = async ({ data, state, set }, slug) => {
       errors: e.response?.data?.errors,
       busy: false,
     });
+
+    await dispatch(displayAlert, {
+      type: "error",
+      text: `There was an error deleting article titled ${title}`,
+    })
   }
 };
 
-export const postComment = async ({ state, set }, { slug, comment }) => {
+export const postComment = async ({ state, set, dispatch }, { slug, comment }) => {
   const { api, article } = state;
 
   try {
@@ -100,7 +148,12 @@ export const postComment = async ({ state, set }, { slug, comment }) => {
     // Refresh comments from the back end at this point
     article.comments = await api.Articles.getComments(slug);
 
-    await set({ article: article });
+    await set({ article });
+
+    await dispatch(displayAlert, {
+      type: "success",
+      text: `Comment added successfully.`,
+    });
   }
   catch (e) {
     await set({
@@ -109,8 +162,33 @@ export const postComment = async ({ state, set }, { slug, comment }) => {
   }
 };
 
-export const deleteComment = async ({ state, set }, { slug, commentId }) => {
-  const { api } = state;
+export const deleteComment = async ({ state, set, dispatch }, { slug, commentId }) => {
+  const { api, article } = state;
+
+  const comment = article.comments.find(c => c.id === commentId);
+  const author = comment.author.username;
+  
+  const yesButton = <button className="btn btn-primary">Yes</button>;
+
+  const result = await dispatch(modal, {
+    title: "Confirm deleting comment",
+    content: (
+      <>
+        <div>Are you sure you want to delete {author}'s comment saying:</div>
+        <div>{comment.body}</div>
+      </>
+    ),
+    buttons: [
+      yesButton,
+      <button className="btn btn-secondary">No!</button>
+    ]
+  });
+
+  if (result !== yesButton) {
+    return;
+  }
+
+  await set({ busy: true });
 
   try {
     await api.Articles.deleteComment(slug, commentId);
@@ -118,11 +196,20 @@ export const deleteComment = async ({ state, set }, { slug, commentId }) => {
     const { article } = state;
     article.comments = await api.Articles.getComments(slug);
 
-    await set({ article });
+    await set({
+      article,
+      busy: false,
+    });
+
+    await dispatch(displayAlert, {
+      type: "success",
+      text: `Successfully removed ${author}'s comment`
+    });
   }
   catch (e) {
     await set({
       errors: e.response?.data?.errors,
+      busy: false,
     });
   }
 };
